backend/requirements.txt
# Phase 1: FastAPI foundation. Phase 2: LlamaIndex + Google GenAI (Gemini) for RAG MVP.
# Pins verified together on Python 3.14 (2026-04); adjust if you standardize on 3.11–3.12.
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.8.1

llama-index-core==0.14.21
llama-index-readers-file==0.6.0
llama-index-embeddings-google-genai==0.5.0
llama-index-llms-google-genai==0.9.1
